requirements.txt
Django==5.0
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
django-cors-headers==4.3.1
djoser==2.2.2
pydantic==2.6.1
